Add `TargetPort` to `RouteToApp` & use it to route connections to multi-port TCP apps (#49047)
* Add TargetPort to RouteToApp and AppMetadata proto messages

* Pass TargetPort during cert generation

* Refactor Pack.makeTLSConfig to accept struct

This will make it easier to add targetPort to it.

* Add labels to UUIDs used by appaccess test pack app servers

This makes them easier to distinguish when routing doesn't work as expected.

* Refactor Pack.CreateAppSession to accept a struct

* TestTCP: Create app session within test

If we kept the old code, we'd need to manually create a session for each
target port, which would create a lot of duplication.

* Prepare integration test fixtures for multi-port tests

* Add api/utils/net.IsPortInRange

* Use TargetPort when routing TCP connections

* Inline dialMultiPortTCPApp, centralize logic

* Check target port when connecting to single-port app

* Reorder check in IsPortInRange

* Use int instead of uint16

* Extract picking dialTarget to separate function

* addr → uriAddr

* Improve err msg for single-port apps when targetPort != uriPort

* Fix unnecessary conversion to int

+++ lib/service/servicecfg/app.go
@@ -137,11 +137,11 @@ type CORS struct {
 // EndPort field is 0.
 type PortRange struct {
 	// Port describes the start of the range. It must be between 1 and 65535.
-	Port uint16
-	// EndPort describes the end of the range, inclusive. If set, it must be between 2and 65535 and be
-	// greater than Port when describing a port range. When omitted or set to zero, it signifies that
-	// the port range defines a single port.
-	EndPort uint16
+	Port int
+	// EndPort describes the end of the range, inclusive. When describing a port range, it must be
+	// greater than Port and less than or equal to 65535. When describing a single port, it must be
+	// set to 0.
+	EndPort int
 }
 
 // CheckAndSetDefaults validates an application.
@@ -224,7 +224,7 @@ func (a *App) checkPorts() error {
 	}
 
 	for _, portRange := range a.TCPPorts {
-		if err := netutils.ValidatePortRange(int(portRange.Port), int(portRange.EndPort)); err != nil {
+		if err := netutils.ValidatePortRange(portRange.Port, portRange.EndPort); err != nil {
 			return trace.Wrap(err, "validating a port range of a TCP app")
 		}
 	}
